// gen/ts/open_splunk/schedule_api.ts
export enum ScheduleValidationMode {
  SCHEDULE_VALIDATION_MODE_UNSPECIFIED = 0,
  SCHEDULE_VALIDATION_MODE_SCHEDULED_REPORT = 1,
  SCHEDULE_VALIDATION_MODE_WEBHOOK_ALERT = 2,
  UNRECOGNIZED = -1,
}
// ...
export enum ScheduleValidationField {
  SCHEDULE_VALIDATION_FIELD_UNSPECIFIED = 0,
  SCHEDULE_VALIDATION_FIELD_CRON = 1,
  SCHEDULE_VALIDATION_FIELD_TIMEZONE = 2,
  SCHEDULE_VALIDATION_FIELD_DISPATCH_TTL = 3,
  SCHEDULE_VALIDATION_FIELD_WEBHOOK_TTL = 4,
  UNRECOGNIZED = -1,
}
// ...
export enum ScheduleValidationCode {
  SCHEDULE_VALIDATION_CODE_UNSPECIFIED = 0,
  SCHEDULE_VALIDATION_CODE_REQUIRED = 1,
  SCHEDULE_VALIDATION_CODE_INVALID = 2,
  SCHEDULE_VALIDATION_CODE_TOO_LARGE = 3,
  UNRECOGNIZED = -1,
}